P4ADEV-1648 move classes
