+++ src/test/java/it/gov/pagopa/pu/worker/mapper/PaymentsReportingMapperTest.java
@@ -1,6 +1,7 @@
 package it.gov.pagopa.pu.worker.mapper;
 
 import it.gov.pagopa.payhub.activities.dto.paymentsreporting.PaymentsReportingDTO;
+import it.gov.pagopa.pu.worker.paymentsreporting.mapper.PaymentsReportingMapper;
 import it.gov.pagopa.pu.worker.paymentsreporting.model.PaymentsReporting;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
